OAK-4180 - Use another NodeStore as a local cache for a remote Document store
Disable support for excluded paths as per current logic once a NodeState is loaded from secondary store it assumes that complete subtree for that path should be present in secondary store.

To support excluded path we would need to check on every child access in DelegatingDocumentNodeState that if child is part of secondary or primary. If part of primary then read call needs to be routed to primary. If required we can support that later

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1750606 13f79535-47bb-0310-9956-ffa450edef68